Initial move of process_x_command into different files
