+++ src/main.rs
@@ -2,31 +2,34 @@ extern crate rpassword;
 
 mod cli;
 mod config;
+mod configuration;
 mod crypto;
 mod database;
+mod environments;
+mod integrations;
 mod lib;
+mod parameters;
+mod projects;
+mod run;
 mod subprocess;
 mod table;
+mod templates;
 
-use crate::cli::{
-    CONFIRM_FLAG, DELETE_SUBCMD, FORMAT_OPT, GET_SUBCMD, LIST_SUBCMD, NAME_ARG, RENAME_OPT,
-    SECRETS_FLAG, SET_SUBCMD, TEMPLATE_FILE_OPT, VALUES_FLAG,
-};
 use crate::config::env::ConfigEnv;
 use crate::config::{Config, CT_PROFILE, DEFAULT_ENV_NAME};
-use crate::database::{
-    Environments, Integrations, OpenApiConfig, ParamExportFormat, ParamExportOptions,
-    ParameterDetails, Parameters, Projects, Templates,
-};
-use crate::subprocess::{EnvSettings, Inheritance, SubProcess};
-use crate::table::Table;
+use crate::configuration::process_config_command;
+use crate::database::{Environments, Integrations, OpenApiConfig, Parameters, Projects, Templates};
+use crate::environments::process_environment_command;
+use crate::integrations::process_integrations_command;
+use crate::parameters::process_parameters_command;
+use crate::projects::process_project_command;
+use crate::run::process_run_command;
+use crate::subprocess::SubProcess;
+use crate::templates::process_templates_command;
 use clap::ArgMatches;
 use color_eyre::eyre::Result;
-use color_eyre::Report;
-use rpassword::read_password;
 use std::io::{self, stdin, stdout, Write};
-use std::str::FromStr;
-use std::{fs, process};
+use std::process;
 use termcolor::{Color, ColorChoice, ColorSpec, StandardStream, WriteColor};
 
 const REDACTED: &str = "*****";
@@ -215,143 +218,6 @@ fn resolve_ids(config: &Config, rest_cfg: &OpenApiConfig) -> Result<ResolvedIds>
     })
 }
 
-/// Process the 'run' sub-command
-fn process_run_command(
-    subcmd_args: &ArgMatches,
-    rest_cfg: &OpenApiConfig,
-    sub_proc: &mut SubProcess,
-    resolved: &ResolvedIds,
-) -> Result<()> {
-    let mut arguments: Vec<String>;
-    let command: String;
-
-    let parameters = Parameters::new();
-    let param_map = parameters.get_parameter_values(
-        rest_cfg,
-        resolved.project_id(),
-        resolved.environment_id(),
-        false,
-    )?;
-    let mut ct_vars = EnvSettings::new();
-    let mut errors: Vec<String> = vec![];
-    for (k, v) in param_map {
-        ct_vars.insert(k.clone(), v.value.clone());
-        if !v.error.is_empty() {
-            errors.push(format_param_error(&k, &v.error))
-        }
-    }
-    sub_proc.set_cloudtruth_environment(ct_vars);
-
-    if subcmd_args.is_present("command") {
-        command = subcmd_args.value_of("command").unwrap().to_string();
-        arguments = vec![];
-    } else if subcmd_args.is_present("arguments") {
-        arguments = subcmd_args.values_of_lossy("arguments").unwrap();
-        command = arguments.remove(0);
-        if command.contains(' ') {
-            warn_user("command contains spaces, and may fail.".to_string())?;
-            let mut reformed = format!("{} {}", command, arguments.join(" "));
-            reformed = reformed.replace("$", "\\$");
-            println!(
-                "Try using 'cloudtruth run --command \"{}\"'",
-                reformed.trim()
-            );
-        }
-    } else {
-        warn_missing_subcommand("run")?;
-        process::exit(0);
-    }
-
-    // NOTE: do this before running the sub-process, since it could be a long-running task
-    warn_unresolved_params(&errors)?;
-
-    // Setup the environment for the sub-process.
-    let inherit = Inheritance::from_str(subcmd_args.value_of("inheritance").unwrap()).unwrap();
-    let overrides = subcmd_args.values_of_lossy("set").unwrap_or_default();
-    let removals = subcmd_args.values_of_lossy("remove").unwrap_or_default();
-    let permissive = subcmd_args.is_present("permissive");
-    sub_proc.set_environment(resolved, inherit, &overrides, &removals)?;
-    if !permissive {
-        sub_proc.remove_ct_app_vars();
-    }
-    sub_proc.run_command(command.as_str(), &arguments)?;
-
-    Ok(())
-}
-
-/// Process the 'project' sub-command
-fn process_project_command(
-    subcmd_args: &ArgMatches,
-    rest_cfg: &OpenApiConfig,
-    projects: &Projects,
-) -> Result<()> {
-    if let Some(subcmd_args) = subcmd_args.subcommand_matches(DELETE_SUBCMD) {
-        let proj_name = subcmd_args.value_of(NAME_ARG).unwrap();
-        let details = projects.get_details_by_name(rest_cfg, proj_name)?;
-
-        if let Some(details) = details {
-            // NOTE: the server is responsible for checking if children exist
-            let mut confirmed = subcmd_args.is_present(CONFIRM_FLAG);
-            if !confirmed {
-                confirmed = user_confirm(format!("Delete project '{}'", proj_name));
-            }
-
-            if !confirmed {
-                warning_message(format!("Project '{}' not deleted!", proj_name))?;
-            } else {
-                projects.delete_project(rest_cfg, &details.id)?;
-                println!("Deleted project '{}'", proj_name);
-            }
-        } else {
-            warning_message(format!("Project '{}' does not exist!", proj_name))?;
-        }
-    } else if let Some(subcmd_args) = subcmd_args.subcommand_matches(LIST_SUBCMD) {
-        let details = projects.get_project_details(rest_cfg)?;
-        if details.is_empty() {
-            println!("No projects found.");
-        } else if !subcmd_args.is_present(VALUES_FLAG) {
-            let list = details
-                .iter()
-                .map(|v| v.name.clone())
-                .collect::<Vec<String>>();
-            println!("{}", list.join("\n"));
-        } else {
-            let fmt = subcmd_args.value_of(FORMAT_OPT).unwrap();
-            let mut table = Table::new("project");
-            table.set_header(&["Name", "Description"]);
-            for entry in details {
-                table.add_row(vec![entry.name, entry.description]);
-            }
-            table.render(fmt)?;
-        }
-    } else if let Some(subcmd_args) = subcmd_args.subcommand_matches(SET_SUBCMD) {
-        let proj_name = subcmd_args.value_of(NAME_ARG).unwrap();
-        let rename = subcmd_args.value_of(RENAME_OPT);
-        let description = subcmd_args.value_of("description");
-        let details = projects.get_details_by_name(rest_cfg, proj_name)?;
-
-        if let Some(details) = details {
-            if description.is_none() && rename.is_none() {
-                warning_message(format!(
-                    "Project '{}' not updated: no updated parameters provided",
-                    proj_name
-                ))?;
-            } else {
-                let name = rename.unwrap_or(&proj_name);
-                projects.update_project(rest_cfg, name, &details.id, description)?;
-                println!("Updated project '{}'", name);
-            }
-        } else {
-            projects.create_project(rest_cfg, proj_name, description)?;
-            println!("Created project '{}'", proj_name);
-        }
-    } else {
-        warn_missing_subcommand("projects")?;
-    }
-
-    Ok(())
-}
-
 /// Process the 'completion' sub-command
 fn process_completion_command(subcmd_args: &ArgMatches) {
     let shell = subcmd_args.value_of("SHELL").unwrap();
@@ -361,749 +227,6 @@ fn process_completion_command(subcmd_args: &ArgMatches) {
         shell.parse().unwrap(),
         &mut io::stdout(),
     );
-}
-
-/// Process the 'config' sub-command
-fn process_config_command(
-    subcmd_args: &ArgMatches,
-    profile_name: Option<&str>,
-    api_key: Option<&str>,
-    proj_name: Option<&str>,
-    env_name: Option<&str>,
-) -> Result<()> {
-    if subcmd_args.subcommand_matches("edit").is_some() {
-        Config::edit()?;
-        let filepath = Config::config_file()
-            .unwrap()
-            .into_os_string()
-            .into_string()
-            .unwrap();
-        println!("Edited {}", filepath);
-    } else if let Some(subcmd_args) = subcmd_args.subcommand_matches(LIST_SUBCMD) {
-        let details = Config::get_profile_details()?;
-        if details.is_empty() {
-            println!("No profiles exist in config.");
-        } else if !subcmd_args.is_present(VALUES_FLAG) {
-            let profile_names: Vec<String> = details.iter().map(|v| v.name.clone()).collect();
-            println!("{}", profile_names.join("\n"));
-        } else {
-            let show_secrets = subcmd_args.is_present(SECRETS_FLAG);
-            let fmt = subcmd_args.value_of(FORMAT_OPT).unwrap();
-            let mut table = Table::new("profile");
-            table.set_header(&["Name", "API", "Environment", "Project", "Description"]);
-            for entry in details {
-                let mut api_value = "".to_string();
-                if let Some(api_key) = entry.api_key {
-                    if show_secrets {
-                        api_value = api_key;
-                    } else if !api_key.is_empty() {
-                        api_value = REDACTED.to_string();
-                    }
-                }
-                table.add_row(vec![
-                    entry.name,
-                    api_value,
-                    entry.environment.unwrap_or_default(),
-                    entry.project.unwrap_or_default(),
-                    entry.description.unwrap_or_default(),
-                ]);
-            }
-            table.render(fmt)?;
-        }
-    } else if let Some(subcmd_args) = subcmd_args.subcommand_matches("current") {
-        let show_secrets = subcmd_args.is_present(SECRETS_FLAG);
-        let show_extended = subcmd_args.is_present("extended");
-        let fmt = subcmd_args.value_of(FORMAT_OPT).unwrap();
-        let values = Config::get_sources(profile_name, api_key, proj_name, env_name)?;
-
-        let mut table = Table::new("profile");
-        table.set_header(&["Parameter", "Value", "Source"]);
-        for v in values {
-            if show_extended || !v.extension {
-                let val_str = if show_secrets || !v.secret || v.value.is_empty() {
-                    v.value
-                } else {
-                    REDACTED.to_string()
-                };
-                table.add_row(vec![v.name, val_str, v.source]);
-            }
-        }
-
-        table.render(fmt)?;
-    } else {
-        warn_missing_subcommand("configuration")?;
-    }
-    Ok(())
-}
-
-/// Process the 'environment' sub-command
-fn process_environment_command(
-    subcmd_args: &ArgMatches,
-    rest_cfg: &OpenApiConfig,
-    environments: &Environments,
-) -> Result<()> {
-    if let Some(subcmd_args) = subcmd_args.subcommand_matches(DELETE_SUBCMD) {
-        let env_name = subcmd_args.value_of(NAME_ARG).unwrap();
-        let details = environments.get_details_by_name(rest_cfg, env_name)?;
-
-        if let Some(details) = details {
-            // NOTE: the server is responsible for checking if children exist
-            let mut confirmed = subcmd_args.is_present(CONFIRM_FLAG);
-            if !confirmed {
-                confirmed = user_confirm(format!("Delete environment '{}'", env_name));
-            }
-
-            if !confirmed {
-                warning_message(format!("Environment '{}' not deleted!", env_name))?;
-            } else {
-                environments.delete_environment(rest_cfg, details.id)?;
-                println!("Deleted environment '{}'", env_name);
-            }
-        } else {
-            warning_message(format!("Environment '{}' does not exist!", env_name))?;
-        }
-    } else if let Some(subcmd_args) = subcmd_args.subcommand_matches(LIST_SUBCMD) {
-        let details = environments.get_environment_details(rest_cfg)?;
-        // NOTE: should always have at least the default environment
-        if !subcmd_args.is_present(VALUES_FLAG) {
-            let list = details
-                .iter()
-                .map(|v| v.name.clone())
-                .collect::<Vec<String>>();
-            println!("{}", list.join("\n"));
-        } else {
-            let fmt = subcmd_args.value_of(FORMAT_OPT).unwrap();
-            let mut table = Table::new("environment");
-            table.set_header(&["Name", "Parent", "Description"]);
-            for entry in details {
-                table.add_row(vec![entry.name, entry.parent_name, entry.description]);
-            }
-            table.render(fmt)?;
-        }
-    } else if let Some(subcmd_args) = subcmd_args.subcommand_matches(SET_SUBCMD) {
-        let env_name = subcmd_args.value_of(NAME_ARG).unwrap();
-        let parent_name = subcmd_args.value_of("parent");
-        let description = subcmd_args.value_of("description");
-        let rename = subcmd_args.value_of(RENAME_OPT);
-        let details = environments.get_details_by_name(rest_cfg, env_name)?;
-
-        if let Some(details) = details {
-            if parent_name.is_some() && parent_name.unwrap() != details.parent_name.as_str() {
-                error_message(format!(
-                    "Environment '{}' parent cannot be updated.",
-                    env_name
-                ))?;
-                process::exit(6);
-            } else if description.is_none() && rename.is_none() {
-                warning_message(format!(
-                    "Environment '{}' not updated: no updated parameters provided",
-                    env_name
-                ))?;
-            } else {
-                let name = rename.unwrap_or(env_name);
-                environments.update_environment(rest_cfg, &details.id, name, description)?;
-                println!("Updated environment '{}'", name);
-            }
-        } else {
-            let parent_name = parent_name.unwrap_or(DEFAULT_ENV_NAME);
-            if let Some(parent_details) = environments.get_details_by_name(rest_cfg, parent_name)? {
-                environments.create_environment(
-                    rest_cfg,
-                    env_name,
-                    description,
-                    parent_details.url.as_str(),
-                )?;
-                println!("Created environment '{}'", env_name);
-            } else {
-                error_message(format!("No parent environment '{}' found", parent_name))?;
-                process::exit(5);
-            }
-        }
-    } else {
-        warn_missing_subcommand("environments")?;
-    }
-    Ok(())
-}
-
-/// Process the 'integrations' sub-command
-fn process_integrations_command(
-    subcmd_args: &ArgMatches,
-    rest_cfg: &OpenApiConfig,
-    integrations: &Integrations,
-) -> Result<()> {
-    if let Some(subcmd_args) = subcmd_args.subcommand_matches("explore") {
-        let fqn = subcmd_args.value_of("FQN");
-        let nodes = integrations.get_integration_nodes(rest_cfg, fqn)?;
-        let indent = "  ";
-        if nodes.is_empty() {
-            if let Some(fqn) = fqn {
-                error_message(format!("Nothing found for FQN '{}'!", fqn))?;
-            } else {
-                error_message("No integrations found.".to_string())?;
-            }
-        } else if !subcmd_args.is_present("values") {
-            for node in nodes {
-                println!("{}", node.name);
-                for key in node.content_keys {
-                    println!("{}{{{{ {} }}}}", indent, key);
-                }
-            }
-        } else {
-            let fmt = subcmd_args.value_of("format").unwrap();
-            let mut table = Table::new("integration");
-            table.set_header(&["Name", "FQN"]);
-            for node in nodes {
-                // add the node itself
-                table.add_row(vec![node.name, node.fqn.clone()]);
-                for key in node.content_keys {
-                    let entry_name = format!("{}{{{{ {} }}}}", indent, key);
-                    table.add_row(vec![entry_name, node.fqn.clone()]);
-                }
-            }
-            table.render(fmt)?;
-        }
-    } else if let Some(subcmd_args) = subcmd_args.subcommand_matches(LIST_SUBCMD) {
-        let details = integrations.get_integration_details(rest_cfg)?;
-        if details.is_empty() {
-            println!("No integrations found");
-        } else if !subcmd_args.is_present(VALUES_FLAG) {
-            let list = details
-                .iter()
-                .map(|d| d.name.clone())
-                .collect::<Vec<String>>();
-            println!("{}", list.join("\n"))
-        } else {
-            let fmt = subcmd_args.value_of(FORMAT_OPT).unwrap();
-            let mut table = Table::new("integration");
-            table.set_header(&["Name", "FQN", "Status", "Updated", "Description"]);
-            for entry in details {
-                table.add_row(vec![
-                    entry.name,
-                    entry.fqn,
-                    entry.status,
-                    entry.status_time,
-                    entry.description,
-                ]);
-            }
-            table.render(fmt)?;
-        }
-    } else {
-        warn_missing_subcommand("integrations")?;
-    }
-    Ok(())
-}
-
-/// Process the 'parameters' sub-command
-fn process_parameters_command(
-    subcmd_args: &ArgMatches,
-    rest_cfg: &OpenApiConfig,
-    parameters: &Parameters,
-    resolved: &ResolvedIds,
-) -> Result<()> {
-    if let Some(subcmd_args) = subcmd_args.subcommand_matches(LIST_SUBCMD) {
-        let proj_id = resolved.project_id();
-        let env_id = resolved.environment_id();
-        let show_secrets = subcmd_args.is_present(SECRETS_FLAG);
-        let mut details =
-            parameters.get_parameter_details(rest_cfg, proj_id, env_id, !show_secrets)?;
-        let references = subcmd_args.is_present("dynamic");
-        let qualifier = if references { "dynamic " } else { "" };
-        if references {
-            // when displaying dynamic parameters, only show the dynamic ones
-            details.retain(|x| x.dynamic)
-        }
-
-        if details.is_empty() {
-            println!(
-                "No {}parameters found in project {}",
-                qualifier,
-                resolved.project_display_name()
-            );
-        } else if !subcmd_args.is_present(VALUES_FLAG) {
-            let list = details
-                .iter()
-                .map(|d| d.key.clone())
-                .collect::<Vec<String>>();
-            println!("{}", list.join("\n"))
-        } else {
-            let fmt = subcmd_args.value_of(FORMAT_OPT).unwrap();
-            let mut errors: Vec<String> = vec![];
-            let mut table = Table::new("parameter");
-
-            if !references {
-                table.set_header(&["Name", "Value", "Source", "Type", "Secret", "Description"]);
-            } else {
-                table.set_header(&["Name", "FQN", "JMES"]);
-            }
-
-            for entry in details {
-                if !entry.error.is_empty() {
-                    errors.push(format_param_error(&entry.key, &entry.error));
-                }
-                if !references {
-                    let type_str = if entry.dynamic { "dynamic" } else { "static" };
-                    let secret_str = if entry.secret { "true" } else { "false" };
-                    table.add_row(vec![
-                        entry.key,
-                        entry.value,
-                        entry.env_name,
-                        type_str.to_string(),
-                        secret_str.to_string(),
-                        entry.description,
-                    ]);
-                } else {
-                    table.add_row(vec![entry.key, entry.fqn, entry.jmes_path]);
-                }
-            }
-            table.render(fmt)?;
-
-            warn_unresolved_params(&errors)?;
-        }
-    } else if let Some(subcmd_args) = subcmd_args.subcommand_matches(GET_SUBCMD) {
-        let key = subcmd_args.value_of("KEY").unwrap();
-        let proj_id = resolved.project_id();
-        let env_id = resolved.environment_id();
-        let parameter = parameters.get_details_by_name(rest_cfg, proj_id, env_id, key, false);
-
-        if let Ok(details) = parameter {
-            // Treat parameters without values set as if the value were simply empty, since
-            // we need to display something sensible to the user.
-            let mut param_value = "".to_string();
-            let mut err_msg = "".to_string();
-            if let Some(param) = details {
-                param_value = param.value;
-                err_msg = param.error;
-            }
-            println!("{}", param_value);
-            if !err_msg.is_empty() {
-                warning_message(err_msg)?;
-            }
-        } else {
-            println!(
-                "The parameter '{}' could not be found in your organization.",
-                key
-            );
-        }
-    } else if let Some(subcmd_args) = subcmd_args.subcommand_matches(SET_SUBCMD) {
-        let key_name = subcmd_args.value_of("KEY").unwrap();
-        let proj_id = resolved.project_id();
-        let env_id = resolved.environment_id();
-        let prompt_user = subcmd_args.is_present("prompt");
-        let filename = subcmd_args.value_of("input-file");
-        let fqn = subcmd_args.value_of("FQN");
-        let jmes_path = subcmd_args.value_of("JMES");
-        let mut value = subcmd_args.value_of("value");
-        let val_str: String;
-        let description = subcmd_args.value_of("description");
-        let rename = subcmd_args.value_of(RENAME_OPT);
-        let final_name = rename.unwrap_or(key_name);
-        let mut param_added = false;
-        let secret: Option<bool> = match subcmd_args.value_of("secret") {
-            Some("false") => Some(false),
-            Some("true") => Some(true),
-            _ => None,
-        };
-
-        // make sure the user did not over-specify
-        if (jmes_path.is_some() || fqn.is_some())
-            && (value.is_some() || prompt_user || filename.is_some())
-        {
-            error_message(
-                concat!(
-                    "Conflicting arguments: cannot specify prompt/input-file/value, ",
-                    "and fqn/jmes-path"
-                )
-                .to_string(),
-            )?;
-            process::exit(7);
-        }
-
-        // if user asked to be prompted
-        if prompt_user {
-            println!("Please enter the '{}' value: ", key_name);
-            val_str = read_password()?;
-            value = Some(val_str.as_str());
-        } else if let Some(filename) = filename {
-            val_str = fs::read_to_string(filename).expect(FILE_READ_ERR);
-            value = Some(val_str.as_str());
-        }
-
-        // make sure there is at least one parameter to updated
-        if description.is_none()
-            && secret.is_none()
-            && value.is_none()
-            && jmes_path.is_none()
-            && fqn.is_none()
-            && rename.is_none()
-        {
-            warn_user(
-                concat!(
-                    "Nothing changed. Please provide at least one of: ",
-                    "description, rename, secret, or value/fqn/jmes-path."
-                )
-                .to_string(),
-            )?;
-        } else {
-            // get the original values, so that is not lost
-            let mut updated: ParameterDetails;
-            if let Some(original) =
-                parameters.get_details_by_name(rest_cfg, proj_id, env_id, key_name, true)?
-            {
-                // only update if there is something to update
-                if description.is_some() || secret.is_some() || rename.is_some() {
-                    updated = parameters.update_parameter(
-                        rest_cfg,
-                        proj_id,
-                        &original.id,
-                        &final_name,
-                        description,
-                        secret,
-                    )?;
-                    // copy a few fields to insure we detect the correct environment
-                    updated.val_id = original.val_id;
-                    updated.env_url = original.env_url;
-                    updated.env_name = original.env_name;
-                } else {
-                    // nothing to update here, but need to copy details
-                    updated = original;
-                }
-            } else {
-                param_added = true;
-                updated = parameters.create_parameter(
-                    rest_cfg,
-                    proj_id,
-                    key_name,
-                    description,
-                    secret,
-                )?;
-            }
-
-            // don't do anything if there's nothing to do
-            if value.is_some() || fqn.is_some() || jmes_path.is_some() {
-                let param_id = updated.id.as_str();
-                // if any existing environment does not match the desired environment
-                if !updated.env_url.contains(env_id) {
-                    let value_add_result = parameters.create_parameter_value(
-                        rest_cfg, proj_id, env_id, param_id, value, fqn, jmes_path,
-                    );
-                    if let Err(err) = value_add_result {
-                        if param_added {
-                            let _ = parameters.delete_parameter_by_id(rest_cfg, proj_id, param_id);
-                        }
-                        return Err(Report::new(err));
-                    }
-                } else {
-                    parameters.update_parameter_value(
-                        rest_cfg,
-                        proj_id,
-                        param_id,
-                        &updated.val_id,
-                        value,
-                        fqn,
-                        jmes_path,
-                    )?;
-                }
-            }
-            println!(
-                "Successfully updated parameter '{}' in project '{}' for environment '{}'.",
-                final_name,
-                resolved.project_display_name(),
-                resolved.environment_display_name(),
-            );
-        }
-    } else if let Some(subcmd_args) = subcmd_args.subcommand_matches(DELETE_SUBCMD) {
-        let key_name = subcmd_args.value_of("KEY").unwrap();
-        let proj_id = resolved.project_id();
-        let env_id = resolved.environment_id();
-        let result = parameters.delete_parameter(rest_cfg, proj_id, env_id, key_name);
-        match result {
-            Ok(Some(_)) => {
-                println!(
-                    "Successfully removed parameter '{}' from project '{}'.",
-                    key_name,
-                    resolved.project_display_name(),
-                );
-            }
-            Ok(None) => {
-                println!(
-                    "Did not find parameter '{}' to delete from project '{}'.",
-                    key_name,
-                    resolved.project_display_name(),
-                )
-            }
-            _ => {
-                println!(
-                    "Failed to remove parameter '{}' from project '{}'.",
-                    key_name,
-                    resolved.project_display_name(),
-                );
-            }
-        };
-    } else if let Some(subcmd_args) = subcmd_args.subcommand_matches("export") {
-        let proj_id = resolved.project_id();
-        let env_id = resolved.environment_id();
-        let starts_with = subcmd_args.value_of("starts_with");
-        let ends_with = subcmd_args.value_of("ends_with");
-        let contains = subcmd_args.value_of("contains");
-        let template_format = subcmd_args.value_of("FORMAT").unwrap();
-        let export = subcmd_args.is_present("export");
-        let secrets = subcmd_args.is_present(SECRETS_FLAG);
-        let options = ParamExportOptions {
-            format: ParamExportFormat::from_str(template_format).unwrap(),
-            starts_with: starts_with.map(|s| s.to_string()),
-            ends_with: ends_with.map(|s| s.to_string()),
-            contains: contains.map(|s| s.to_string()),
-            export: Some(export),
-            secrets: Some(secrets),
-        };
-        let body = parameters.export_parameters(rest_cfg, proj_id, env_id, options)?;
-
-        if let Some(body) = body {
-            println!("{}", body)
-        } else {
-            println!(
-                "Could not export parameters format '{}' from project '{}' in environment '{}'.",
-                template_format,
-                resolved.project_display_name(),
-                resolved.environment_display_name()
-            )
-        }
-    } else if let Some(subcmd_args) = subcmd_args.subcommand_matches("unset") {
-        let key_name = subcmd_args.value_of("KEY").unwrap();
-        let proj_id = resolved.project_id();
-        let env_id = resolved.environment_id();
-        let result = parameters.delete_parameter_value(rest_cfg, proj_id, env_id, key_name);
-        match result {
-            Ok(Some(_)) => {
-                println!(
-                    "Successfully removed parameter value '{}' from project '{}' for environment '{}'.",
-                    key_name,
-                    resolved.project_display_name(),
-                    resolved.environment_display_name()
-                );
-            }
-            Ok(None) => {
-                println!(
-                    "Did not find parameter value '{}' to delete from project '{}' for environment '{}'.",
-                    key_name,
-                    resolved.project_display_name(),
-                    resolved.environment_display_name()
-                )
-            }
-            _ => {
-                println!(
-                    "Failed to remove parameter value '{}' from project '{}' for environment '{}'.",
-                    key_name,
-                    resolved.project_display_name(),
-                    resolved.environment_display_name()
-                );
-            }
-        };
-    } else if let Some(subcmd_args) = subcmd_args.subcommand_matches("differences") {
-        let show_secrets = subcmd_args.is_present(SECRETS_FLAG);
-        let fmt = subcmd_args.value_of(FORMAT_OPT).unwrap();
-        let properties: Vec<&str> = subcmd_args.values_of("properties").unwrap().collect();
-        let env1_name = subcmd_args.value_of("ENV1").unwrap();
-        let env2_name = subcmd_args.value_of("ENV2").unwrap();
-
-        if env1_name == env2_name {
-            warning_message("Invalid comparing an environment to itself".to_string())?;
-        } else {
-            let proj_id = resolved.project_id();
-
-            // fetch all environments once, and then determine id's from the same map that is
-            // used to resolve the environment names.
-            let environments = Environments::new();
-            let env_url_map = environments.get_url_name_map(rest_cfg);
-            let env1_id = environments.id_from_map(env1_name, &env_url_map)?;
-            let env2_id = environments.id_from_map(env2_name, &env_url_map)?;
-
-            let env1_values = parameters.get_parameter_detail_map(
-                rest_cfg,
-                &env_url_map,
-                proj_id,
-                &env1_id,
-                !show_secrets,
-            )?;
-            let env2_values = parameters.get_parameter_detail_map(
-                rest_cfg,
-                &env_url_map,
-                proj_id,
-                &env2_id,
-                !show_secrets,
-            )?;
-            let mut param_list: Vec<String> = env1_values.iter().map(|(k, _)| k.clone()).collect();
-            param_list.sort_by_key(|l| l.to_lowercase());
-
-            let default_param = ParameterDetails::default();
-            let mut added = false;
-            let mut table = Table::new("parameter");
-            let mut errors: Vec<String> = vec![];
-            table.set_header(&["Parameter", env1_name, env2_name]);
-            for param_name in param_list {
-                let details1 = env1_values.get(&param_name).unwrap_or(&default_param);
-                let details2 = env2_values.get(&param_name).unwrap_or(&default_param);
-                let env1 = details1.get_properties(&properties).join(",\n");
-                let env2 = details2.get_properties(&properties).join(",\n");
-                if !details1.error.is_empty() {
-                    errors.push(format_param_error(&param_name, &details1.error))
-                }
-                // NOTE: do not put redundant errors on the list, but the errors could be due to
-                //       different FQNs
-                if !details2.error.is_empty() && details1.error != details2.error {
-                    errors.push(format_param_error(&param_name, &details2.error))
-                }
-                if env1 != env2 {
-                    table.add_row(vec![param_name, env1, env2]);
-                    added = true;
-                }
-            }
-            if added {
-                table.render(fmt)?;
-            } else {
-                println!("No parameters or differences in compared properties found.");
-            }
-            warn_unresolved_params(&errors)?;
-        }
-    } else {
-        warn_missing_subcommand("parameters")?;
-    }
-    Ok(())
-}
-
-/// Process the 'templates' sub-command
-fn process_templates_command(
-    subcmd_args: &ArgMatches,
-    rest_cfg: &OpenApiConfig,
-    templates: &Templates,
-    resolved: &ResolvedIds,
-) -> Result<()> {
-    if let Some(subcmd_args) = subcmd_args.subcommand_matches(DELETE_SUBCMD) {
-        let proj_name = resolved.project_display_name();
-        let proj_id = resolved.project_id();
-        let template_name = subcmd_args.value_of(NAME_ARG).unwrap();
-        let details = templates.get_details_by_name(rest_cfg, proj_id, template_name)?;
-
-        if let Some(details) = details {
-            let mut confirmed = subcmd_args.is_present(CONFIRM_FLAG);
-            if !confirmed {
-                confirmed = user_confirm(format!(
-                    "Delete template '{}' in project '{}'",
-                    template_name, proj_name
-                ));
-            }
-
-            if !confirmed {
-                warning_message(format!(
-                    "Template '{}' in project '{}' not deleted!",
-                    template_name, proj_name
-                ))?;
-            } else {
-                templates.delete_template(rest_cfg, proj_id, &details.id)?;
-                println!(
-                    "Deleted template '{}' in project '{}'",
-                    template_name, proj_name
-                );
-            }
-        } else {
-            warning_message(format!(
-                "Template '{}' does not exist for project '{}'!",
-                template_name, proj_name
-            ))?;
-        }
-    } else if let Some(subcmd_args) = subcmd_args.subcommand_matches(LIST_SUBCMD) {
-        let proj_name = resolved.project_display_name();
-        let proj_id = resolved.project_id();
-        let details = templates.get_template_details(rest_cfg, proj_id)?;
-        if details.is_empty() {
-            println!("No templates in project '{}'.", proj_name);
-        } else if !subcmd_args.is_present(VALUES_FLAG) {
-            let list = details
-                .iter()
-                .map(|n| n.name.clone())
-                .collect::<Vec<String>>();
-            println!("{}", list.join("\n"))
-        } else {
-            let fmt = subcmd_args.value_of(FORMAT_OPT).unwrap();
-            let mut table = Table::new("template");
-            table.set_header(&["Name", "Description"]);
-            for entry in details {
-                table.add_row(vec![entry.name, entry.description]);
-            }
-            table.render(fmt)?;
-        }
-    } else if let Some(subcmd_args) = subcmd_args.subcommand_matches(GET_SUBCMD) {
-        let proj_name = resolved.project_display_name();
-        let proj_id = resolved.project_id();
-        let env_id = resolved.environment_id();
-        let template_name = subcmd_args.value_of(NAME_ARG).unwrap();
-        let show_secrets = subcmd_args.is_present(SECRETS_FLAG);
-        let body =
-            templates.get_body_by_name(rest_cfg, proj_id, env_id, template_name, show_secrets)?;
-
-        if let Some(body) = body {
-            println!("{}", body)
-        } else {
-            error_message(format!(
-                "No template '{}' found in project '{}'.",
-                template_name, proj_name
-            ))?;
-            process::exit(9);
-        }
-    } else if let Some(subcmd_args) = subcmd_args.subcommand_matches("preview") {
-        let proj_id = resolved.project_id();
-        let env_id = resolved.environment_id();
-        let show_secrets = subcmd_args.is_present(SECRETS_FLAG);
-        let filename = subcmd_args.value_of(TEMPLATE_FILE_OPT).unwrap();
-        let body = fs::read_to_string(filename).expect(FILE_READ_ERR);
-        let result = templates.preview_template(rest_cfg, proj_id, env_id, &body, show_secrets)?;
-        println!("{}", result);
-    } else if let Some(subcmd_args) = subcmd_args.subcommand_matches(SET_SUBCMD) {
-        let filename = subcmd_args.value_of(TEMPLATE_FILE_OPT);
-        let proj_id = resolved.project_id();
-        let proj_name = resolved.project_display_name();
-        let template_name = subcmd_args.value_of(NAME_ARG).unwrap();
-        let rename = subcmd_args.value_of(RENAME_OPT);
-        let description = subcmd_args.value_of("description");
-        let details = templates.get_details_by_name(rest_cfg, proj_id, template_name)?;
-
-        if let Some(details) = details {
-            if description.is_none() && rename.is_none() && filename.is_none() {
-                warning_message(format!(
-                    "Template '{}' not updated: no updated parameters provided",
-                    template_name
-                ))?;
-            } else {
-                let name = rename.unwrap_or(&template_name);
-                let mut body = None;
-                if let Some(filename) = filename {
-                    body = Some(fs::read_to_string(filename).expect(FILE_READ_ERR));
-                }
-                templates.update_template(
-                    rest_cfg,
-                    proj_id,
-                    &details.id,
-                    name,
-                    description,
-                    body.as_deref(),
-                )?;
-                println!("Updated template '{}' in project '{}'", name, proj_name);
-            }
-        } else if let Some(filename) = filename {
-            let body = fs::read_to_string(filename).expect(FILE_READ_ERR);
-            templates.create_template(rest_cfg, proj_id, template_name, &body, description)?;
-            println!(
-                "Created template '{}' in project '{}'",
-                template_name, proj_name
-            );
-        } else {
-            error_message("Must provide a body for a new template".to_owned())?;
-            process::exit(8);
-        }
-    } else {
-        warn_missing_subcommand("templates")?;
-    }
-    Ok(())
 }
 
 fn main() -> Result<()> {
